[feat/api]: Les endpoints de API sont fonctionnelles.

+++ src/main/java/com/deep_coding15/GesStockApi/stock/dto/stock_mouvement/StockMouvementResponseDTO.java
@@ -7,8 +7,8 @@ import lombok.Setter;
 @Setter
 public class StockMouvementResponseDTO {
     public Long id;
-    public int produitId;
-    public int quantite;
+    public Long produitId;
+    public Integer quantite;
     public String type;
     public Long utilisateurId;
     public String dateMouvement;
